Promote `Mesh` indices from `u16` to `u32`
  - Also fixes a crash when rendering big meshes in OpenGL.

+++ src/graphics/backend_wgpu/mod.rs
@@ -122,7 +122,7 @@ impl Gpu {
     pub(super) fn draw_triangles(
         &mut self,
         vertices: &[Vertex],
-        indices: &[u16],
+        indices: &[u32],
         view: &TargetView,
         transformation: &Transformation,
     ) {
